Initial attempt to use tanks engines to start apps, while still using bcore's wrapper.
It takes a long time to initialize an engine, which might be due to the shotgun connection
that it creates. On the other hand, there should be a way to rely solely on paths on disk
to do that.

Now I have the tendency to implement the application specific part myself, all with the
location of the multi-launch app on disk to be able to use their particular startup scripts.